+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/docs
 
+## 1.9.2
+
+### Patch Changes
+
+- @inglorious/engine@9.0.0
+- @inglorious/renderer-2d@8.0.0
+- @inglorious/renderer-react-dom@8.0.0
+
 ## 1.9.1
 
 ### Patch Changes