+++ examples/games/flapper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # inglorious-flapper
 
+## 0.2.2
+
+### Patch Changes
+
+- @inglorious/engine@9.0.0
+- @inglorious/renderer-2d@8.0.0
+
 ## 0.2.1
 
 ### Patch Changes