+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/engine
 
+## 9.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@6.0.0
+
 ## 8.0.0
 
 ### Patch Changes