+++ packages/renderer-2d/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/renderer-2d
 
+## 8.0.0
+
+### Patch Changes
+
+- @inglorious/engine@9.0.0
+
 ## 7.0.0
 
 ### Patch Changes