+++ packages/server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/server
 
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@6.0.0
+
 ## 6.0.0
 
 ### Patch Changes